Fix dates on blogposts
Hugo had done some update that made it not interpret a string as HTML.
This was fixed in the Papermod theme, but bumping the theme caused other
collisions with my modifications, so for now just add this as another
modification.
https://github.com/adityatelange/hugo-PaperMod/commit/7795c90f6fa106733267481d9ed0518b00f2c62c#diff-a6fb449eb5948ec17cb5af3700d5b678f045a2edf4330889d7bbb7ad61cfefc7

Note to self: using "latest version" is not a good idea.

Signed-off-by: Casper Andersson <casper.casan@gmail.com>